Created subscriber and added documentation
