+++ bundles/org.palladiosimulator.analyzer.slingshot.eventdriver/src/org/palladiosimulator/analyzer/slingshot/eventdriver/entity/Subscriber.java
@@ -14,6 +14,8 @@ import org.palladiosimulator.analyzer.slingshot.eventdriver.returntypes.Result;
 import io.reactivex.rxjava3.disposables.Disposable;
 import io.reactivex.rxjava3.functions.Consumer;
 
+import org.apache.log4j.Logger;
+
 /**
  * A subscriber to an event of type {@code T} that should be activated upon the event.
  * It holds a reference to a {@link EventHandler}.
@@ -35,6 +37,8 @@ import io.reactivex.rxjava3.functions.Consumer;
  * @param <T> The event type
  */
 public class Subscriber<T> implements Consumer<T>, Disposable, Comparable<Subscriber<T>>, InterceptorInformation {
+		
+	private static final Logger LOGGER = Logger.getLogger(Subscriber.class);
 
 	private boolean disposed = false;
 
@@ -47,7 +51,7 @@ public class Subscriber<T> implements Consumer<T>, Disposable, Comparable<Subscr
 	private final Class<T> forEvent;
 	private Optional<IPreInterceptor> preInterceptor;
 	private Optional<IPostInterceptor> postInterceptor;
-	private final List<EventContract> associatedContracts;
+	private final List<SubscriberContract> associatedContracts;
 
 	private Subscriber(final Builder<T> builder) {
 		this.priority = builder.priority;
@@ -81,7 +85,8 @@ public class Subscriber<T> implements Consumer<T>, Disposable, Comparable<Subscr
 		final InterceptionResult postInterceptionResult = this.postInterceptor
 				.map(postInterceptor -> postInterceptor.apply(preInterceptionInformation, event, result))
 				.orElseGet(InterceptionResult::success);
-
+		
+		LOGGER.info("Post interception result was successful: " + postInterceptionResult.wasSuccessful());
 	}
 
 	@Override
@@ -98,7 +103,7 @@ public class Subscriber<T> implements Consumer<T>, Disposable, Comparable<Subscr
 
 	@Override
 	public int compareTo(final Subscriber<T> other) {
-		if (this.getEventType().equals(other.getEnclosingType())) {
+		if (this.getEventType().equals(other.getEventType())) {
 			return Integer.compare(priority, other.priority);
 		}
 
@@ -161,7 +166,7 @@ public class Subscriber<T> implements Consumer<T>, Disposable, Comparable<Subscr
 	}
 
 	@Override
-	public List<EventContract> getAssociatedContracts() {
+	public List<SubscriberContract> getAssociatedContracts() {
 		return associatedContracts;
 	}
 
@@ -197,7 +202,7 @@ public class Subscriber<T> implements Consumer<T>, Disposable, Comparable<Subscr
 		private final Class<T> forEvent;
 		private Optional<IPreInterceptor> preInterceptor = Optional.empty();
 		private Optional<IPostInterceptor> postInterceptor = Optional.empty();
-		private List<EventContract> associatedContracts = Collections.emptyList();
+		private List<SubscriberContract> associatedContracts = Collections.emptyList();
 
 		private Builder(final Class<T> forEvent) {
 			this.forEvent = forEvent;
@@ -243,7 +248,7 @@ public class Subscriber<T> implements Consumer<T>, Disposable, Comparable<Subscr
 			return this;
 		}
 
-		public Builder<T> associatedContracts(final List<EventContract> associatedContracts) {
+		public Builder<T> associatedContracts(final List<SubscriberContract> associatedContracts) {
 			this.associatedContracts = associatedContracts;
 			return this;
 		}
